Extract redirects from InfoMessageFilterFeature into a separate feature
