+++ common/src/main/resources/assets/wynntils/lang/en_us.json
@@ -32,6 +32,31 @@
   "feature.wynntils.bombBellTracking.name": "Bomb Bell Tracking",
   "feature.wynntils.chatCoordinates.name": "Chat Coordinates",
   "feature.wynntils.chatItem.name": "Chat Item",
+  "feature.wynntils.chatRedirect.craftedDurability.description": "How should crafted item durability messages appear?",
+  "feature.wynntils.chatRedirect.craftedDurability.name": "Crafted Durability Filtering",
+  "feature.wynntils.chatRedirect.friendJoin.description": "How should friend join messages appear?",
+  "feature.wynntils.chatRedirect.friendJoin.name": "Friend Join Message Filtering",
+  "feature.wynntils.chatRedirect.heal.description": "How should messages about healing appear?",
+  "feature.wynntils.chatRedirect.heal.name": "Heal Message Filtering",
+  "feature.wynntils.chatRedirect.horse.description": "How should messages relating to your horse appear?",
+  "feature.wynntils.chatRedirect.horse.name": "Horse Message Filtering",
+  "feature.wynntils.chatRedirect.loginAnnouncements.description": "How should login messages appear?",
+  "feature.wynntils.chatRedirect.loginAnnouncements.name": "Login Message Filtering",
+  "feature.wynntils.chatRedirect.name": "Info Message Filter",
+  "feature.wynntils.chatRedirect.notEnoughMana.description": "How should messages about not enough mana when casting appear?",
+  "feature.wynntils.chatRedirect.notEnoughMana.name": "Mana Message Filtering",
+  "feature.wynntils.chatRedirect.potion.description": "How should messages relating to your health potion stack appear?",
+  "feature.wynntils.chatRedirect.potion.name": "Potion Stack Message Filtering",
+  "feature.wynntils.chatRedirect.shaman.description": "How should shaman-specific messages appear?",
+  "feature.wynntils.chatRedirect.shaman.name": "Shaman Message Filtering",
+  "feature.wynntils.chatRedirect.soulPoint.description": "How should soul point messages appear?",
+  "feature.wynntils.chatRedirect.soulPoint.name": "Soul Point Message Filtering",
+  "feature.wynntils.chatRedirect.speed.description": "How should messages about speed effect appear?",
+  "feature.wynntils.chatRedirect.speed.name": "Speed Effect Message Filtering",
+  "feature.wynntils.chatRedirect.toolDurability.description": "How should gathering tool durability messages appear?",
+  "feature.wynntils.chatRedirect.toolDurability.name": "Gather Tool Durability Filtering",
+  "feature.wynntils.chatRedirect.unusedPoints.description": "How should messages about unused ability or skill points appear?",
+  "feature.wynntils.chatRedirect.unusedPoints.name": "Unused Ability and Skill Point Message Filtering",
   "feature.wynntils.chatTabs.name": "Chat Tabs",
   "feature.wynntils.chatTimestamp.formatPattern.description": "How should the timestamp look like in chat?",
   "feature.wynntils.chatTimestamp.formatPattern.name": "Timestamp Format",
@@ -178,37 +203,13 @@
   "feature.wynntils.infoBox.overlay.infoBox.secondsPerRecalculation.name": "Info Box Content Seconds Per Recalculation",
   "feature.wynntils.infoBox.overlay.infoBox.textShadow.description": "Should the info overlay text be drawn with a shadow?",
   "feature.wynntils.infoBox.overlay.infoBox.textShadow.name": "Text Shadow",
-  "feature.wynntils.infoMessageFilter.craftedDurability.description": "How should crafted item durability messages appear?",
-  "feature.wynntils.infoMessageFilter.craftedDurability.name": "Crafted Durability Filtering",
-  "feature.wynntils.infoMessageFilter.friendJoin.description": "How should friend join messages appear?",
-  "feature.wynntils.infoMessageFilter.friendJoin.name": "Friend Join Message Filtering",
-  "feature.wynntils.infoMessageFilter.heal.description": "How should messages about healing appear?",
-  "feature.wynntils.infoMessageFilter.heal.name": "Heal Message Filtering",
   "feature.wynntils.infoMessageFilter.hideLevelUp.description": "Should level up messages be hidden?",
   "feature.wynntils.infoMessageFilter.hideLevelUp.name": "Hide Level Up Messages",
   "feature.wynntils.infoMessageFilter.hideSystemInfo.description": "Should Wynncraft info messages be hidden?",
   "feature.wynntils.infoMessageFilter.hideSystemInfo.name": "Hide Info Messages",
   "feature.wynntils.infoMessageFilter.hideWelcome.description": "Should Wynncraft welcome messages be hidden?",
   "feature.wynntils.infoMessageFilter.hideWelcome.name": "Hide Welcome Messages",
-  "feature.wynntils.infoMessageFilter.horse.description": "How should messages relating to your horse appear?",
-  "feature.wynntils.infoMessageFilter.horse.name": "Hide Horse Messages",
-  "feature.wynntils.infoMessageFilter.loginAnnouncements.description": "How should login messages appear?",
-  "feature.wynntils.infoMessageFilter.loginAnnouncements.name": "Login Message Filtering",
   "feature.wynntils.infoMessageFilter.name": "Info Message Filter",
-  "feature.wynntils.infoMessageFilter.notEnoughMana.description": "How should messages about not enough mana when casting appear?",
-  "feature.wynntils.infoMessageFilter.notEnoughMana.name": "Mana Message Filtering",
-  "feature.wynntils.infoMessageFilter.potion.description": "How should messages relating to your health potion stack appear?",
-  "feature.wynntils.infoMessageFilter.potion.name": "Potion Stack Message Filtering",
-  "feature.wynntils.infoMessageFilter.shaman.description": "How should shaman-specific messages appear?",
-  "feature.wynntils.infoMessageFilter.shaman.name": "Shaman Message Filtering",
-  "feature.wynntils.infoMessageFilter.soulPoint.description": "How should soul point messages appear?",
-  "feature.wynntils.infoMessageFilter.soulPoint.name": "Soul Point Message Filtering",
-  "feature.wynntils.infoMessageFilter.speed.description": "How should messages about speed effect appear?",
-  "feature.wynntils.infoMessageFilter.speed.name": "Speed Effect Message Filtering",
-  "feature.wynntils.infoMessageFilter.toolDurability.description": "How should gathering tool durability messages appear?",
-  "feature.wynntils.infoMessageFilter.toolDurability.name": "Gather Tool Durability Filtering",
-  "feature.wynntils.infoMessageFilter.unusedPoints.description": "How should messages about unused ability or skill points appear?",
-  "feature.wynntils.infoMessageFilter.unusedPoints.name": "Unused Ability and Skill Point Message Filtering",
   "feature.wynntils.ingredientPouchHotkey.name": "Ingredient Pouch Keybind",
   "feature.wynntils.inventoryEmeraldCount.emeraldCountType.description": "How should emerald count be displayed?",
   "feature.wynntils.inventoryEmeraldCount.emeraldCountType.name": "Emerald Count Render Type",
